added framework for generic small tasks thread pool that has yet to be verified against another model (i.e. gpu)
